requirements.txt
numpy >= 1.19.0, < 2.0.0
scipy >= 1.0.1
future >= 0.16.0
matplotlib >= 3.0.0
pillow >= 6.2.0
opencv-python >= 4.6.0
scikit-image >= 0.14.0
PyYAML
torch
torchvision
imageio
